refactor(core): own backend device printing
